Abstracted API logic from vuex to services
