+++ store/twitter.js
@@ -1,12 +1,4 @@
-import Twit from 'twit';
-
-const T = new Twit({
-  consumer_key:         process.env.consumerKey,
-  consumer_secret:      process.env.consumerSecret,
-  access_token:         process.env.accessToken,
-	access_token_secret:  process.env.accessTokenSecret,
-	timeout_ms:           60 * 1000
-});
+import {getTimeline} from '~/services/TwitterService';
 
 export const state = () => ({
 	timeline: []
@@ -20,6 +12,6 @@ export const mutations = {
 
 export const actions = {
 	async getTimeline(store) {
-		T.get('statuses/user_timeline', { screen_name: 'vuejsmcr', count: 9 }).then(result =>	store.commit('buildTimeline', result.data));
+		await getTimeline().then(result => store.commit('buildTimeline', result.data));
 	}
 };
